removed some things from git

+++ hSD/env/point_environment.py
@@ -21,8 +21,8 @@ class PointEnv(py_environment.PyEnvironment):
 
     self.box_size, self.step_size = box_size, step_size
     self.start_state = 0., 0.
-    self._action_spec = array_spec.BoundedArraySpec(shape=(2,), dtype=np.float64, minimum=-step_size, maximum=step_size, name='action')
-    self._observation_spec = array_spec.BoundedArraySpec(shape=(2,), dtype=np.float64, minimum=-box_size, maximum=box_size, name='observation')
+    self._action_spec = array_spec.BoundedArraySpec(shape=(2,), dtype=np.float32, minimum=-step_size, maximum=step_size, name='action')
+    self._observation_spec = array_spec.BoundedArraySpec(shape=(2,), dtype=np.float32, minimum=-box_size, maximum=box_size, name='observation')
     self._state = self.start_state
     self._step_count = 0
     self._episode_ended = False
@@ -37,7 +37,7 @@ class PointEnv(py_environment.PyEnvironment):
     self._state = self.start_state
     self._step_count = 0
     self._episode_ended = False
-    return ts.restart(np.array(self._state, dtype=np.float64))
+    return ts.restart(np.array(self._state, dtype=np.float32))
 
   def _clip_to_bounds(self, val, low, high):
     return min(high, max(val, low))
@@ -55,9 +55,9 @@ class PointEnv(py_environment.PyEnvironment):
 
     if self._step_count > 1000:
       self._episode_ended = True
-      return ts.termination(np.array(self._state, dtype=np.float64), reward=0)
+      return ts.termination(np.array(self._state, dtype=np.float32), reward=0)
     else:
-      return ts.transition(np.array(self._state, dtype=np.float64), reward=0)
+      return ts.transition(np.array(self._state, dtype=np.float32), reward=0)
 
   def set_start_state(self, state):
     self.start_state = state
